+++ CHANGELOG.md
@@ -6,6 +6,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.7] - 2022-12-02
+### Added
+- Added a `file_checksum_method` config option that can be used to select whether to use file content hashing (default)
+or file modification timestamps for calculating the checksum of an external file represented by a `Path` object
+
+### Fixed
+- Fixed a bug where CLI arguments containing a hyphen would not be parsed correctly when used in a `Lab`
+- Fixed a bug where the implementation of coverage in `labfile.py` would break the debugger for the entire script. Now
+only the `coverage` steps breaks debugging (which is expected, since coverage replaces the tracing function used by the
+debugger)
+
+### Changed
+- Made the `call()` utility function print program outputs (stdout) to stdout by default - this can be controlled using
+the `echo_output_to_stream` argument.
+
 ## [0.0.6] - 2022-05-17
 ### Added
 - Added py.typed file to signal to downstream user's that alkymi has type annotations
@@ -117,7 +132,8 @@ from cache
 ### Added
 - Initial release
 
-[Unreleased]: https://github.com/MathiasStokholm/alkymi/compare/v0.0.6...HEAD
+[Unreleased]: https://github.com/MathiasStokholm/alkymi/compare/v0.0.7...HEAD
+[0.0.7]: https://github.com/MathiasStokholm/alkymi/compare/v0.0.6...v0.0.7
 [0.0.6]: https://github.com/MathiasStokholm/alkymi/compare/v0.0.5...v0.0.6
 [0.0.5]: https://github.com/MathiasStokholm/alkymi/compare/v0.0.4...v0.0.5
 [0.0.4]: https://github.com/MathiasStokholm/alkymi/compare/v0.0.3...v0.0.4